backend/requirements.txt
Flask==3.0.3
tensorflow==2.17.0
Pillow==10.3.0
numpy==1.26.4
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.1.0
flask-cors==5.0.0